+++ linksmith/sphinx/core.py
@@ -8,8 +8,30 @@ import requests
 
 from linksmith.model import OutputFormat, OutputFormatRegistry, ResourceType
 from linksmith.sphinx.inventory import InventoryFormatter
+from linksmith.sphinx.util import LocalObjectsInv
 
 logger = logging.getLogger(__name__)
+
+
+def dump_inventory_universal(infiles: t.List[str], format_: str = "text"):
+    """
+    Decode one or multiple intersphinx inventories and output in different formats.
+    """
+    if not infiles:
+        logger.info("No inventory specified, entering auto-discovery mode")
+        try:
+            local_objects_inv = LocalObjectsInv.discover(Path.cwd())
+            logger.info(f"Auto-discovered objects.inv: {local_objects_inv}")
+            infiles = [str(local_objects_inv)]
+        except Exception as ex:
+            raise FileNotFoundError(f"No inventory specified, and none discovered: {ex}")
+    for infile in infiles:
+        if infile.endswith(".inv"):
+            inventory_to_text(infile, format_=format_)
+        elif infile.endswith(".txt"):
+            inventories_to_text(infile, format_=format_)
+        else:
+            raise NotImplementedError(f"Unknown input file type: {infile}")
 
 
 def inventory_to_text(url: str, format_: str = "text"):